refactor: move ShutdownCather into util
Only expose CompressedWriter which is just a writer with a finish fn that
takes care of the unwrapping etc.
